Go back to int64 for Size, and failing test.

+++ cache/cache_store.go
@@ -9,7 +9,7 @@ import (
 // on what BinaryStore is being used.
 type Metadata struct {
 	ContentType string
-	Size uint
+	Size int64
 }
 
 type MetadataStore interface {
